requirements.txt
whitenoise==5.0.1
Django==2.2.12
django-crispy-forms==1.8.1
django-filter==2.2.0
gunicorn==20.0.4
sqlparse==0.3.0
django-import-export==2.1.0

